% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bind_na.R
\name{cbindna}
\alias{cbindna}
\title{qpcR cbind.na method.}
\usage{
cbindna(..., deparse.level = 1)
}
\arguments{
\item{...}{(generalized) vectors or matrices. See base::cbind}

\item{deparse.level}{integer controlling the construction
of labels in the case of non-matrix-like arguments. See base::cbind}
}
\value{
a matrix combining the ... arguments column-wise.
}
\description{
qpcR cbind.na method.
}
\examples{
### Vectors:
a <- c(rep(1, 5), NA, seq(1:5))
b <- c(rep(1, 4), NA, seq(1:7))

# Complete shorter vector with NAs:
cbindna(a,b)

}
