% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bind_na.R
\name{rbindna}
\alias{rbindna}
\title{qpcR rbind.na method.}
\usage{
rbindna(..., deparse.level = 1)
}
\arguments{
\item{...}{(generalized) vectors or matrices. See base::rbind}

\item{deparse.level}{integer controlling the construction
of labels in the case of non-matrix-like arguments. See base::rbind}
}
\value{
a matrix combining the ... arguments row-wise.
}
\description{
qpcR rbind.na method.
}
\examples{
### Vectors:
a <- c(rep(1, 5), NA, seq(1:5))
b <- c(rep(1, 4), NA, seq(1:7))

# Complete shorter vector with NAs:
rbindna(a,b)

}
